layout and basic functions
