+++ console/templates/layouts/_base.html
@@ -45,6 +45,7 @@
     <script src="https://code.jquery.com/jquery-3.1.1.js"
             integrity="sha256-16cdPddA6VdVInumRGo6IbivbERE8p7CQR3HzTBuELA="
             crossorigin="anonymous"></script>
+    <script src="/static/js/app.js?v=18"></script>
 
   </head>
