+++ console/templates/partials/navigation_tabs.html
@@ -1,38 +1,38 @@
 <ul class="navigation-tabs navigation-tabs--pills">
 
     <li id="Home"
-        class="navigation-tabs__item navigation-tabs__item--primary {% if label == 'Home' %}{{'navigation-tabs__item--active'}}{% endif %}">
+        class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/home' %} navigation-tabs__item--active {% endif %}">
         <a href="/home"
-           class="navigation-tabs__tab {% if label == 'Home' %}{{'navigation-tabs__tab--active'}}{% endif %}"> Home </a>
+           class="btn btn--primary btn--border navigation-tabs__tab"> Home </a>
     </li>
 
     <li id="Decrypt"
-        class="navigation-tabs__item navigation-tabs__item--primary {% if label == 'Decrypt' %}{{'navigation-tabs__item--active'}}{% endif %}">
+        class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/decrypt' %} navigation-tabs__item--active {% endif %}">
         <a href="/decrypt"
-           class="navigation-tabs__tab {% if label == 'Decrypt' %}{{'navigation-tabs__tab--active'}}{% endif %}"> Decrypt </a>
+           class="btn btn--primary btn--border navigation-tabs__tab"> Decrypt </a>
     </li>
 
     <li id="Submit"
-        class="navigation-tabs__item navigation-tabs__item--primary {% if label == 'Submit' %}{{'navigation-tabs__item--active'}}{% endif %}">
+        class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/submit' %} navigation-tabs__item--active {% endif %}">
         <a href="/submit"
-           class="navigation-tabs__tab {% if label == 'Submit' %}{{'navigation-tabs__tab--active'}}{% endif %}"> Submit </a>
+           class="btn btn--primary btn--border navigation-tabs__tab"> Submit </a>
     </li>
 
     <li id="Store"
-        class="navigation-tabs__item navigation-tabs__item--primary {% if label == 'Store' %}{{'navigation-tabs__item--active'}}{% endif %}">
+        class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/store' %} navigation-tabs__item--active {% endif %}">
         <a href="/store"
-           class="navigation-tabs__tab {% if label == 'Store' %}{{'navigation-tabs__tab--active'}}{% endif %}"> Store </a>
+           class="navigation-tabs__tab"> Store </a>
     </li>
 
     <li id="Add User"
-        class="navigation-tabs__item navigation-tabs__item--primary {% if label == 'Add User' %}{{'navigation-tabs__item--active'}}{% endif %}">
+        class="navigation-tabs__item navigation-tabs__item--primary {% if request.path == '/add_user' %} navigation-tabs__item--active {% endif %}">
         <a href="/add_user"
-           class="navigation-tabs__tab {% if label == 'Add User' %}{{'navigation-tabs__tab--active'}}{% endif %}"> Add User </a>
+           class="navigation-tabs__tab"> Add User </a>
     </li>
 
     <li id="Logout"
         class="navigation-tabs__item navigation-tabs__item--primary {% if label == 'Logout' %}{{'navigation-tabs__item--active'}}{% endif %}">
         <a href="/home/logout"
-           class="navigation-tabs__tab {% if label == 'Logout' %}{{'navigation-tabs__tab--active'}}{% endif %}"> Logout </a>
+           class="navigation-tabs__tab"> Logout </a>
     </li>
 </ul>
