+++ console/templates/partials/receiver.html
@@ -1,0 +1,84 @@
+<div id='response-data-wrapper' class='col-md-6'>
+      <h3>Receiver <img src="static/images/ajax-loader.gif" id="ftp-loading-sign" style="position: relative; display: none;"></h3>
+      <p>Lists of the <strong>most recent 10 files</strong> to hit the FTP server:</p>
+
+      <!--<ul class='nav nav-tabs' role='tablist'>-->
+        <!--<li role="presentation" class="active"><a href="#pcks" aria-controls="pcks" role="tab" data-toggle="tab">PCKs:</a></li>-->
+        <!--<li role="presentation"><a href="#images" aria-controls="images" role="tab" data-toggle="tab">Images:</a></li>-->
+        <!--<li role="presentation"><a href="#indices" aria-controls="indices" role="tab" data-toggle="tab">Image Indices:</a></li>-->
+        <!--<li role="presentation"><a href="#receipts" aria-controls="receipts" role="tab" data-toggle="tab">Receipts:</a></li>-->
+      <!--</ul>-->
+
+      <ul class="navigation-tabs navigation-tabs--pills" role="tablist">
+
+        <li id="PCK:" class="navigation-tabs__item navigation-tabs__item--primary">
+          <a href="#pcks" class="navigation-tabs__tab ">PCKs</a> </li>
+
+        <li id="Images:" class="navigation-tabs__item navigation-tabs__item--primary">
+          <a href="#images" class="navigation-tabs__tab ">Images</a> </li>
+
+        <li id="Image Indices:" class="navigation-tabs__item navigation-tabs__item--primary">
+          <a href="#indices" class="navigation-tabs__tab ">Image Indices</a> </li>
+
+        <li id="Receipts:" class="navigation-tabs__item navigation-tabs__item--primary">
+          <a href="#receipts" class="navigation-tabs__tab ">Receipts</a> </li>
+
+      </ul>
+
+      <ul>
+        <li class="grid card">
+
+        <div class="grid__col col-2@m">
+            <span class="survey-list-item__label venus">Filename: </span>
+        </div>
+
+        <div class="grid__col col-2@m">
+            <span class="survey-list-item__label venus">Size: </span>
+        </div>
+
+        <div class="grid__col col-2@m">
+            <span class="survey-list-item__label venus">Modified: </span>
+        </div>
+
+        </li>
+      </ul>
+
+
+        <div class="tab-content" id="ftp-tab-content">
+          <div role="tabpanel" class="tab-pane active" id="pcks">
+            <table id="pck-data" class="table table-hover">
+              <tbody>
+                <thead>
+                  <tb>
+                    elo
+                  </tb>
+                  <tb>
+                    hi
+                  </tb>
+                </thead>
+              </tbody>
+            </table>
+          </div>
+
+        <div role="tabpanel" class="tab-pane" id="images">
+          <table id="image-data" class="table table-hover">
+              <tbody>
+              </tbody>
+          </table>
+        </div>
+
+        <div role="tabpanel" class="tab-pane" id="indices">
+          <table id="index-data" class="table table-hover">
+              <tbody>
+              </tbody>
+          </table>
+        </div>
+
+        <div role="tabpanel" class="tab-pane" id="receipts">
+          <table id="receipt-data" class="table table-hover">
+              <tbody>
+              </tbody>
+          </table>
+        </div>
+
+    </div>
